+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Updated react-i18next to 11.15.6
 - Updated i18next to 21.6.13
 - Updated i18next-icu to 2.0.3
+- Updated sidekiq-unique-jobs to 7.1.15
 - Fix a bug: a sentence was not linked to a translation key
 - Fix a security issue: updated image_processing to 1.12.2 to fix [CVE-2022-24720](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-24720)
 - Fix a security issue: updated url-parse to 1.5.10 to fix [CVE-2022-0686](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-0686), [CVE-2022-0691](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-0691), [CVE-2022-0639](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-0639) and [CVE-2022-0512](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-0512)